added error messages

+++ hotel-service/src/main/java/com/example/hotel_service/repository/HotelRepository.java
@@ -3,6 +3,6 @@ package com.example.hotel_service.repository;
 import com.example.hotel_service.entity.Hotel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HotelRepository extends JpaRepository<Hotel, Long> {
+public interface HotelRepository extends JpaRepository<Hotel, Integer> {
 
 }
